week8 assignment markdown
i had trouble with DAPC output BayEnv2 in general.

I coudn't remember how you removed the putatively selected loci from your vcf file, so i finagled my own way using sed. It assumes that the row order is the same in the vcf and bayescan land, so hopefully it is.